resolveWorkspaceUsageDates helper
